day_91_image_color_extractor/requirements.txt
flask
extcolors
webcolors
pillow
plotly
pandas
gunicorn
